myapp/.ipynb_checkpoints/requirements-checkpoint.txt
androguard
beautifulsoup4
cn_clip
cx_Freeze
Markdown
matplotlib
numpy
opencv_python_headless
Pillow
playwright
PySide6
PySide6_Addons
PySide6_Essentials
pyzbar
Requests
scikit_learn
torch
tqdm
zxing
